refactor: address PR #93 review suggestions (round 2)
- Move PortStatusResult into packages/shared/src/types/network.ts and
  export from @showstack/shared — eliminates the duplicate definition
  that existed in both PortStatusMonitorService (main) and
  PortUsageIndicator (renderer)
- Remove redundant else block from network status useEffect — cleanup
  function in the return already handles interval teardown on every
  dependency change and unmount
