+++ packages/shared/src/index.ts
@@ -17,5 +17,6 @@ export * from './validation';
 // Auth utilities
 export * from './auth/passwordPolicy';
 
-// Network utilities
+// Network utilities and types
 export * from './utils/networkValidation';
+export * from './types/network';
